Refactor `Conversation` component
* Create `conversation` folder in `components` to contain all
`Conversation` component related files
* Move all handlers to `handlers.js`
* Move all constants to `constants.js`
* Move all cursor-related functions to `cursor.js`
* Move all generic functions to `utils.js`
* Update router to use correct path for `Conversation` component and
remove the name for the `/conversations` route
