+++ src/router/index.js
@@ -1,7 +1,7 @@
 import Vue from 'vue';
 import VueRouter from 'vue-router';
 import Conversations from '../views/Conversations.vue';
-import Conversation from '../components/Conversation.vue';
+import Conversation from '../components/conversation/Conversation.vue';
 import ConversationsHome from '../components/ConversationsHome.vue';
 
 Vue.use(VueRouter);
@@ -9,7 +9,6 @@ Vue.use(VueRouter);
 const routes = [
   {
     path: '/conversations',
-    name: 'Conversations',
     component: Conversations,
     children: [
       {
